scripts/cstr_enmpc.py: Finish implementation.
